Refactor SetupJob action routines out into a service
Simplify SetupJobAction by moving directory creation, dependencies download and environment evaluation out into a separate service.
